rename main_chain_follower to data_sources in node
